pyproject.toml
[tool.poetry]
name = "home-automation"
version = "1.0"
description = ""
authors = ["Michael Moore"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pycoingecko = "^3.1.0"
influxdb = "^5.3.1"
paho-mqtt = "^1.6.1"
requests = "^2.31.0"
lxml = "^4.9.3"
types-requests = "^2.31.0.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true